requirements.txt
PmlyYAML>=5.4.1
pydantic>=1.8.2
psycopg2-binary>=2.9.3
SQLAlchemy>=1.4.39
tenacity>=8.0.1
numpy>=1.21.6
pandas>=1.3.5
aiohttp>=3.8.1
yfinance>=0.1.70
prometheus-client>=0.14.1
psutil>=5.9.0
python-dateutil>=2.8.2
pytz>=2022.1
requests>=2.27.1
websocket-client>=1.3.3
redis>=4.3.4
mlflow>=1.26.1
onnx>=1.11.0
onnxruntime>=1.11.1
scikit-learn>=1.0.2
matplotlib>=3.5.2
seaborn>=0.11.2
xgboost>=1.5.0
shap>=0.40.0
pytest>=7.0.1
pytest-cov>=3.0.0
black>=22.3.0
flake8>=4.0.1
mypy>=0.950
isort>=5.10.1
pre-commit>=2.19.0
schedule>=1.1.0
keras-tuner>=1.1.0

# NLP and Sentiment Analysis
nltk>=3.7.0
transformers>=4.18.0
torch>=1.11.0
praw>=7.6.0
textblob>=0.17.1
spacy>=3.3.0
pandas-ta>=0.3.14b0
polygon-api-client>=1.6.0

# FinBERT specific
sentencepiece>=0.1.96
